requirements/local.txt
# Local development dependencies go here
-r base.txt

django-debug-toolbar==1.7

pytest==3.0.7
pytest-django==3.1.2
coverage==4.3.4
prospector==0.12.5
unidiff==0.5.3
pylint==1.7.1
